replace uint variable types with byte/unsigned int

+++ LEDBlinker.h
@@ -13,12 +13,13 @@
 #define LED_BLINKER_MIN_INTERVAL 20
 
 // Example setup/running code:
-//    LEDBlinker led(13);  // Use pin 13
+//    LEDBlinker led(13);  // Use LED on pin 13
 //    void setup() {
 //      pinMode(13, OUTPUT);  // Must setup pin as output
 //    }
 //    void loop() { 
 //      led.update(); // Must be called in loop.  Do not block (use delays) in loop
+//      // Do other stuff, call LED control code below to control LED
 //    }
 //
 // Example control code:
@@ -49,12 +50,12 @@
 
 class LEDBlinker {
 private:
-  uint8_t mPin;
-  uint8_t mNumBlinks;
-  uint16_t mBlinkOnDuration; 
-  uint16_t mBlinkOffDuration;
-  uint16_t mPauseDuration;
-  uint16_t mNumSequences;  // 0 to run forever
+  byte mPin;
+  byte mNumBlinks;
+  unsigned int mBlinkOnDuration; 
+  unsigned int mBlinkOffDuration;
+  unsigned int mPauseDuration;
+  unsigned int mNumSequences;  // 0 to run forever
   unsigned long mStartTime;
   unsigned long mLastRunTime;
   void (*mCallbackFunction)();
@@ -64,10 +65,10 @@ protected:
   
 public:
   // Create a new LED Fader for a pin
-  LEDBlinker(uint8_t const pin): mPin(pin), mNumBlinks(0){};
+  LEDBlinker(byte const pin): mPin(pin), mNumBlinks(0){};
   
   // Restarts at first sequence (if one was running) on new pin
-  void setPin(uint8_t const pin);
+  void setPin(byte const pin);
   
   // This starts a sequence of LED blinking, followed by a pause
   // (with the LED off), and repeats the sequence a specified number of times,
@@ -87,21 +88,21 @@ public:
   //      can specify 0 to indicate infinite sequences.
   //   callbackFunction - a function you want to have run when all sequences have
   //      completed.  Ignored if numSequences is 0 (infinite sequences).
-  void blink(uint16_t const blinkOnDuration,
-             uint16_t const blinkOffDuration,
-             uint8_t const numBlinks,
-             uint16_t const pauseDuration,
-             uint16_t const numSequences,
+  void blink(unsigned int const blinkOnDuration,
+             unsigned int const blinkOffDuration,
+             byte const numBlinks,
+             unsigned int const pauseDuration,
+             unsigned int const numSequences,
              void (*callbackFunction)());
   
   // Turn on, leave on for duration, then call callback
-  void onForDuration(uint16_t const onDuration,
+  void onForDuration(unsigned int const onDuration,
                      void (*callbackFunction)()) {
     blink(onDuration, 0, 1, 0, 1, callbackFunction);        
   }
   
   // Turn off, leave off for duration, then call callback
-  void offForDuration(uint16_t const offDuration,
+  void offForDuration(unsigned int const offDuration,
                       void (*callbackFunction)()) {
     blink(0, offDuration, 1, 0, 1, callbackFunction);        
   }
